Work with db

+++ app/src/main/java/com/steelcrow/androchat/chatRoom/OnChatItemClickListener.java
@@ -1,5 +1,5 @@
 package com.steelcrow.androchat.chatRoom;
 
 public interface OnChatItemClickListener {
-    void onItemClick(CharSequence chatTitle, int chatId);
+    void onItemClick(CharSequence chatTitle, long chatId);
 }
